+++ functional-tests/http/src/test/java/org/mule/test/http/functional/requester/streaming/HttpStreamingRequestNtlmProxyConfigTestCase.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.test.http.functional.requester.proxy;
+package org.mule.test.http.functional.requester.streaming;
 
 import static org.mule.runtime.http.api.HttpConstants.HttpStatus.OK;
 import static org.mule.runtime.http.api.HttpHeaders.Names.CONNECTION;
